requirements.txt
requests
rich
click
PyPDF2
PyYAML
streamlit
pytest
